third_party/blink/web_tests/fast/hidpi/image-srcset-ignore-double-descriptor.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
    addEventListener("load", function() {
        shouldBeFalse('document.getElementById("foo").clientWidth==100');
        shouldBeTrue('document.getElementById("foo").clientWidth==200');
    }, false);
</script>
<div>This test passes if the image below says 2x with a greenish background even when the deviceScaleFactor is 1.</div>
<img id="foo" src="" srcset="resources/blue-100-px-square.png 1x 1x, resources/deleteButton.png 3x, resources/green-400-px-square.png 2x">
